DeepTald_back_end/requirements.txt
Flask-Cors==5.0.0
Flask==3.0.3
Cerberus==1.3.5
huggingface_hub==0.25.1
inflection==0.5.1
pytest==8.3.3
python-dotenv==1.0.1
sorcery==0.2.2
tqdm==4.66.5
transformers==4.45.1
ffmpeg-python==0.2.0
openai-whisper @ git+https://github.com/openai/whisper.git@6dea21fd7f7253bfe450f1e2512a0fe47ee2d258
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
sentence-transformers==3.1.1
torch==2.3.1 
torchvision==0.18.1 
torchaudio==2.3.1
torch-audiomentations
torch-metrics
torch-pitch-shift
torchmetrics
pydub
nltk==3.9.1
phonemizer==3.3.0
matplotlib
numpy==1.26.3